% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swpc.R
\name{swpc_alerts}
\alias{swpc_alerts}
\title{Fetch NOAA space weather alerts}
\usage{
swpc_alerts()
}
\value{
tibble: product_id (character), issue_datetime (POSIXct),
message (character)
}
\description{
Returns recent space weather alerts, watches, and warnings issued
by the Space Weather Prediction Center.
}
